create downloadBlobToMachine function, organize files
