+++ src/components/analysis/uploadPart/uploadPartButton.jsx
